Split TkOptimizerContext because that file is huge
